add ui/ux

+++ routes/collectionsRouter.js
@@ -16,11 +16,65 @@ router.delete('/delete', auth, async (req, res) => {
             return res.status(400).json({ message: 'Nothing to delete' })
         }
 
-        const collection = await Collection.findByIdAndDelete(id)
+        const collection = await Collection.findById(id)
+
+        if(!collection) {
+            return res.json({ message: 'Collection do not exist' })
+        }
+
+        await Collection.findByIdAndDelete(id)
 
         await User.findByIdAndUpdate(collection.owner, {$pull: {
             collections: id
         }})
+
+        res.json(collection)
+    } catch (error) {
+        res.status(500).json({ error })
+    }
+})
+
+router.post('/edit', auth, async (req, res) => {
+    try {
+        const { id, topic, title, description, author, comments, year, picture } = req.body
+
+        const collection = await Collection.findById(id)
+
+
+        if(!collection) {
+            return res.status(404).json({ message: `Collection with id ${id} is nonexist` })
+        }
+
+        if(!title){
+            return res.status(400).json({ message: `Title can not be empty` })
+        }
+
+        if(!topic){
+            return res.status(400).json({ message: `Topic can not be empty` })
+        }
+
+        if(collection.owner != req.user.id && req.user.userRole !== 'admin') {
+            return res.status(401).json({ message: 'You do not have permission for this operation' })
+        }
+
+        const { url } = collection.picture !== picture 
+        && picture 
+        && await cloudinary.uploader.upload(picture, {
+            upload_preset: 'ml_default'
+        }, (err, result) => result)
+
+        collection.author = author
+        collection.title = title
+        collection.description = description
+        collection.topic = topic
+        collection.comments = comments
+        collection.year = year
+
+        if(url) {
+            collection.picture = picture
+        }
+
+        await collection.save()
 
         res.json(collection)
     } catch (error) {
@@ -63,13 +117,13 @@ router.get('/', async (req, res) => {
         const { id } = req.query
 
         if(!id){
-            return res.status(404).json({ message: 'This user dont have any collection yet' })
+            return res.status(404).json({ message: 'Id is not correct' })
         }
 
         const collections = await Collection.find({ owner: id})
 
         if(!collections.length){
-            return res.status(404).json({ message: 'This user dont have any collection yet' })
+            return res.json({ message: 'This user dont have any collection yet' })
         }
         
         res.json(collections)
